requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymongo==4.7.3
motor==3.4.0
beanie==1.26.0
pytest==8.2.0
pytest-asyncio==0.25.0
httpx==0.27.0
python-dotenv==1.0.1
langchain>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-mongodb>=0.2.0
langchain-huggingface>=0.0.3
sentence-transformers>=3.0.0
torch>=2.0.0
apscheduler>=3.10.0,<4.0.0
